+++ modules/service/query-service.js
@@ -116,8 +116,7 @@ class QueryService {
 
         const documentPath = this.fileService.getHandlerIdDocumentPath(handlerId);
         const handlerData = await this.fileService.loadJsonFromFile(documentPath);
-        console.log(request);
-        console.log(response);
+
         for (const object of response) {
             const assertion = handlerData.find((x) => x.id === object.assertionId);
             if (assertion) {
